frist

+++ main_async2.py
@@ -189,7 +189,7 @@ def extract_urls_from_folder():
                         if count % 1000 == 0:
                             print(count, datetime.datetime.now())
 
-                        if count < 609000:
+                        if count < 620000:
                             continue
                         if not is_url_exists(conn, url):
                             yield url
@@ -364,7 +364,8 @@ async def main():
 
 
 if __name__ == "__main__":
-    shutil.rmtree(os.path.join(os.getcwd(), f"user_data"))
+    if os.path.isdir(os.path.join(os.getcwd(), f"user_data")):
+        shutil.rmtree(os.path.join(os.getcwd(), f"user_data"))
     asyncio.run(main())
 # 21:02 - 48000
 # 21:10 - 49000
